# Improve performance of hashing of S3 files