(extracted from https://github.com/quiltdata/quilt/pull/1788 since this gives a large speed-up)

Chunk size is increased from 1KB to 256KB.
This allows GIL to be released during hash_obj.update() call.
See https://docs.python.org/3/library/hashlib.html#hashlib.hash.update.

Benchmarked on m5a.large and Python 3.6 (`max_workers` of [`ThreadPoolExecutor`](https://docs.python.org/3/library/concurrent.futures.html#concurrent.futures.ThreadPoolExecutor) defaults to 10), that's the number of files processed concurrently here: 
https://github.com/quiltdata/quilt/blob/4594c1c261e05b9f128d04467a254d9260e77c88/api/python/quilt3/data_transfer.py#L912-L917

Before:

```
In [1]: import quilt3; quilt3.data_transfer.calculate_sha256([quilt3.util.PhysicalKey('ai2-datasets',
                                                              'arc/ARC-V1-Feb2018.zip', None)], [680841265])
Hashing: 100%|███████████████████████████████████████████████████████████████| 681M/681M [00:31<00:00, 21.4MB/s]
Out[1]: ['6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0']

In [1]: import quilt3; quilt3.data_transfer.calculate_sha256([quilt3.util.PhysicalKey('ai2-datasets',
                                                              'arc/ARC-V1-Feb2018.zip', None)]*2, [680841265]*2)
Hashing: 100%|█████████████████████████████████████████████████████████████| 1.36G/1.36G [00:36<00:00, 36.9MB/s]
Out[1]:
['6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0']

In [2]: import quilt3; quilt3.data_transfer.calculate_sha256([quilt3.util.PhysicalKey('ai2-datasets',
                                                              'arc/ARC-V1-Feb2018.zip', None)]*4, [680841265]*4)
Hashing: 100%|█████████████████████████████████████████████████████████████| 2.72G/2.72G [02:35<00:00, 17.5MB/s]
Out[2]:
['6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0']

In [2]: import quilt3; quilt3.data_transfer.calculate_sha256([quilt3.util.PhysicalKey('ai2-datasets',
                                                              'arc/ARC-V1-Feb2018.zip', None)]*10, [680841265]*10)
Hashing: 100%|█████████████████████████████████████████████████████████████| 6.81G/6.81G [07:23<00:00, 15.4MB/s]
Out[2]:
['6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0']
```

After
```
In [1]: import quilt3; quilt3.data_transfer.calculate_sha256([quilt3.util.PhysicalKey('ai2-datasets', 
                                                              'arc/ARC-V1-Feb2018.zip', None)], [680841265])
Hashing: 100%|███████████████████████████████████████████████████████████████| 681M/681M [00:31<00:00, 21.9MB/s]
Out[1]: ['6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0']


In [2]: import quilt3; quilt3.data_transfer.calculate_sha256([quilt3.util.PhysicalKey('ai2-datasets', 
                                                              'arc/ARC-V1-Feb2018.zip', None)]*2, [680841265]*2)
Hashing: 100%|█████████████████████████████████████████████████████████████| 1.36G/1.36G [00:35<00:00, 38.4MB/s]
Out[2]: 
['6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0']

In [3]: import quilt3; quilt3.data_transfer.calculate_sha256([quilt3.util.PhysicalKey('ai2-datasets', 
                                                              'arc/ARC-V1-Feb2018.zip', None)]*4, [680841265]*4)
Hashing: 100%|█████████████████████████████████████████████████████████████| 2.72G/2.72G [00:34<00:00, 79.1MB/s]
Out[3]: 
['6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0']


In [1]: import quilt3; quilt3.data_transfer.calculate_sha256([quilt3.util.PhysicalKey('ai2-datasets', 
                                                              'arc/ARC-V1-Feb2018.zip', None)]*10, [680841265]*10)
Hashing: 100%|██████████████████████████████████████████████████████████████| 6.81G/6.81G [00:36<00:00, 189MB/s]
Out[1]: 
['6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0',
 '6d2d5ab50b2ceec6ba5f79c921be77cf2de712ea25a2b3f4fff3acc101cecfa0']
```