@sir-sigurd 2020-09-22
@akarve could you post your benchmarks? Is that slowdown reproducible?